Adapt remaining callers to be compatible with std::span
After this lands, we need only test that our clients can build
with the std::span alias, after which we can remove the flag.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1104616
Commit-Queue: Mike Reed <mike@reedtribe.org>

+++ bench/PathOpsBench.cpp
@@ -191,8 +191,8 @@ protected:
             case MakeType::kArray: {
             //    ArrayPath<N*12> arrays;
             //    run_builder(arrays, false, N);
-                return SkPath::Raw({fArrays.fPts, fArrays.fPIndex},
-                                   {fArrays.fVbs, fArrays.fVIndex},
+                return SkPath::Raw({fArrays.fPts, (size_t)fArrays.fPIndex},
+                                   {fArrays.fVbs, (size_t)fArrays.fVIndex},
                                    {}, SkPathFillType::kWinding);
             }
         }
