+++ fuzz/FuzzCanvasHelpers.cpp
@@ -405,7 +405,7 @@ static sk_sp<SkPathEffect> make_fuzz_patheffect(Fuzz* fuzz, int depth) {
             int count;
             fuzz->nextRange(&count, 0, (int)std::size(intervals));
             fuzz->nextN(intervals, count);
-            return SkDashPathEffect::Make({intervals, count}, phase);
+            return SkDashPathEffect::Make({intervals, (size_t)count}, phase);
         }
         case 8: {
             SkScalar segLength, dev;
@@ -963,19 +963,22 @@ static sk_sp<SkTextBlob> make_fuzz_textblob(Fuzz* fuzz) {
                 fuzz->next(&x, &y);
                 // TODO: Test other variations of this.
                 buffer = &textBlobBuilder.allocRun(font, glyphCount, x, y);
-                (void)font.textToGlyphs(textPtr, textLen, encoding, {buffer->glyphs, glyphCount});
+                (void)font.textToGlyphs(textPtr, textLen, encoding,
+                                        {buffer->glyphs, (size_t)glyphCount});
                 break;
             case 1:
                 fuzz->next(&y);
                 // TODO: Test other variations of this.
                 buffer = &textBlobBuilder.allocRunPosH(font, glyphCount, y);
-                (void)font.textToGlyphs(textPtr, textLen, encoding, {buffer->glyphs, glyphCount});
+                (void)font.textToGlyphs(textPtr, textLen, encoding,
+                                        {buffer->glyphs, (size_t)glyphCount});
                 fuzz->nextN(buffer->pos, glyphCount);
                 break;
             case 2:
                 // TODO: Test other variations of this.
                 buffer = &textBlobBuilder.allocRunPos(font, glyphCount);
-                (void)font.textToGlyphs(textPtr, textLen, encoding, {buffer->glyphs, glyphCount});
+                (void)font.textToGlyphs(textPtr, textLen, encoding,
+                                        {buffer->glyphs, (size_t)glyphCount});
                 fuzz->nextN(buffer->pos, glyphCount * 2);
                 break;
             default:
