+++ tests/RectTest.cpp
@@ -139,7 +139,7 @@ DEF_TEST(Rect_setbounds, reporter) {
     SkRect r;
     const SkRect zeror = { 0, 0, 0, 0 };
     for (const SkPoint* pts : { p0, p1, p2, p3 }) {
-        for (int n = 1; n <= 4; ++n) {
+        for (size_t n = 1; n <= 4; ++n) {
             bool isfinite = r.setBoundsCheck({pts, n});
             REPORTER_ASSERT(reporter, !isfinite);
             REPORTER_ASSERT(reporter, r == zeror);
